cast (int) in coremap

+++ os161-base-2.0.3/kern/vm/coremap.c
@@ -108,7 +108,7 @@ static paddr_t get_n_frames(unsigned int num, bool fixed, pt_entry* entry) {
         return 0;
     } else if (!found) {
         i = get_victim();
-        if(i == -1){
+        if((int)i == -1){
             lock_release(coremap_lock);
             return 0;
         }
